Widen libbitcoinkernel's block header nonce to 64 bits for ProgPoW
btck_block_header_get_nonce() (and the C++ wrapper's Nonce()) still
returned uint32_t and read the old CBlockHeader::nNonce field, which
this fork replaced with a 64-bit nNonce64 for ProgPoW. Truncating would
silently drop real nonce bits for any consumer of this API, so widened
the return type throughout instead. Dormant either way (BUILD_KERNEL_LIB
is off), verified by building the bitcoinkernel target standalone.

+++ bitcoin-source/src/kernel/bitcoinkernel.h
@@ -2024,12 +2024,14 @@ BITCOINKERNEL_API int32_t btck_block_header_get_version(
     const btck_BlockHeader* header) BITCOINKERNEL_ARG_NONNULL(1);
 
 /**
- * @brief Get the nonce from btck_BlockHeader.
+ * @brief Get the nonce from btck_BlockHeader. GPU-mined-crypto note: this is
+ * ProgPoW's 64-bit nonce (CBlockHeader::nNonce64), not upstream's original
+ * 32-bit SHA-256d nonce.
  *
  * @param[in] header    Non-null btck_BlockHeader
  * @return              Nonce
  */
-BITCOINKERNEL_API uint32_t btck_block_header_get_nonce(
+BITCOINKERNEL_API uint64_t btck_block_header_get_nonce(
     const btck_BlockHeader* header) BITCOINKERNEL_ARG_NONNULL(1);
 
 /**
